pyproject.toml
[tool.poetry]
name = "lezgian_tts"
version = "0.1.0"
description = "Project for Lezgian TTS"
authors = ["Vazhid Gadzhiev <vazhid64@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
flask = "^3.1"
scipy = "^1.15.2"
flask-cors = "^6.0.0"
torch = "^2.6.0"
transformers = "^4.51.3"
gunicorn = "^23.0.0"
soundfile = "^0.13.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
flask-login = ">=0.5.0"
werkzeug = "^3.1.3"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
locust = "^2.37.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
